Refactor: add util directory
- 유효성 관련 함수를 signup에서 분리하여 다른곳에서도 재사용이 가능하도록
구현합니다.
